+++ bindgen/src/bindings/cpp/templates/err_conv.hpp
@@ -4,5 +4,5 @@ struct {{ typ|ffi_converter_name }} {
     static RustBuffer lower(const {{ class_name }} &);
     static {{ type_name }} read(RustStream &stream);
     static void write(RustStream &stream, const {{ class_name }} &);
-    static int32_t allocation_size(const {{ class_name }} &);
+    static uint64_t allocation_size(const {{ class_name }} &);
 };
